Move report_simulation_errors into DriverLogger
